Update main branch to update staging / production environments.

+++ services/hanke-service/src/test/kotlin/fi/hel/haitaton/hanke/domain/HankeTest.kt
@@ -11,7 +11,7 @@ internal class HankeTest {
 
     @ParameterizedTest(name = "There are {0} days between {1} and {2}")
     @CsvSource("1,2021-03-02,2021-03-02", "214,2021-03-01,2021-09-30")
-    fun haittaAjanKesto(expectedNumberOfDays: Long, startDate: LocalDate, endDate: LocalDate) {
+    fun haittaAjanKesto(expectedNumberOfDays: Int, startDate: LocalDate, endDate: LocalDate) {
         val hanke = Hanke(1, "HAI21-1")
         hanke.haittaAlkuPvm = startDate.atStartOfDay(TZ_UTC)
         hanke.haittaLoppuPvm = endDate.atStartOfDay(TZ_UTC)
